backend/requirements.txt
Flask
python-docx
google-generativeai
python-dotenv
Flask-Cors